image/requirements.txt
beautifulsoup4==4.12.3
boto3==1.35.39
botocore==1.35.39
docxcompose==1.4.0
fastapi==0.115.2
fastembed==0.3.6
fuzzywuzzy==0.18.0
groq==0.11.0
jsonschema==4.23.0
langchain==0.3.3
langchain_community==0.3.2
langchain_text_splitters==0.3.0
langdetect==1.0.9
lxml==5.3.0
mangum==0.19.0
numpy==2.1.2
openai==1.51.2
pandas==2.2.3
protobuf==5.28.2
pydantic==2.9.2
PyPDF2==3.0.1
pytest==8.3.3
python-dotenv==1.0.1
python_docx==1.1.2
PyYAML==6.0.2
PyYAML==6.0.2
qdrant_client==1.12.0
Requests==2.32.3
scikit_learn==1.5.2
tqdm==4.66.5
underthesea==6.8.4
uvicorn==0.31.1
